musl: Don't add cheri-clang to dynamic linker name
We have picked up a change which results in the dynamic linker
being called ld-musl-riscv64cheri-clang.so.1, while we don't have
a corresponding change to the name of the dynamic linker required by
executables. While it might be nice to adopt this, for the time being
we don't need it, so the simple fix is to remove it.