eudev: Fix hashmap assertions
The HashmapBase structure has interesting alignmnent tricks that
won't work with the CHERI requirement that pointers and the structs
that contain them are size aligned. Add alignment attributes that
silence the compiler.

These additional alignment requirements and the increased size of
pointers in structures force us to artificially increase the direct
storage size by one byte.

The strange division performed by DIRECT_BUCKETS (adding one byte
to the entry structure size) is fine because the actual entries
are put a the start of the direct storage area followed by an
array of bytes (actually dib_raw_t) with the same number of entries.