codasip-cheri.bbclass: Use eudev as device manager not busybox
It seems that the busybox implementation of udev does not
support any type of hotplug or device events. It just creates
the device nodes. This results in X11 input device not being
recognized by the X-server by default.

Now that the kernel will no longer raise CHERI faults when trying
to run udevd for this purpose switch to udevd for device node
management.