gtk+3: Add fixes for CHERI _get_type issue
Update the template used for generating _get_type() functions so that
they use g_once_init_{enter,leave}_pointer where appropriate.

We also need to update the PACK and UNPACK macros used with the
GtkCssNodeDeclaration type. It stores additional information in the
bottom bits of the pointer, but we need to ensure the macros use
GTYPE_TO_POINTER and GPOINTER_TO_TYPE instead of the SIZE equivalent,
as this uses uintptr_t as the intermediate integer type which ensures
that the pointers don't loose provenance.

Signed-off-by: Stuart Menefy <stuart.menefy@codasip.com>