bash: fix cast of SIMPLE_COM to COMMAND
BASH makes the assumption the 'redirects' field is at the same offset
in SIMPLE_COM and COMMAND. This is only the case if int, enum, and
pointers are of equal size. With CHERI, however, this is not true and,
for example, "echo foo |& tee bar" would cause a CHERI fault.

Fix this by aligning the declartion order of the shared members of the
SIMPLE_COM and COMMAND structs.