ltp: Add check for virtualization to run script
LTP allows a test to check if it runs in a virtualized environment, e.g.
for a test to increase a timeout due to the overhead. However, these
builtin LTP virtualization checks do not notice if we run inside Qemu.

One way to tell LTP about virtualization is to use the environment
variable 'LTP_VIRT_OVERRIDE'. We add a heuristic to check if there are
virtIO devices present in our system and if so, define
LTP_VIRT_OVERRIDE. While in theory it is possible to run Qemu without
virtIO device models, in practice most people will want to use them.

Signed-off-by: Chris Hofer <christian.hofer@codasip.com>