kmod: Fix alignment calculation
The alignment fix in upstream code relies (intentionally or
by accident) on the fact that the value before padding is already
aligned to 4 bytes. As a pointer can be at most 8 bytes in size
the padding required is always equal to the misalignment of the
padded value (either 0 or 4).

For CHERI possible misalignments are 0, 4, 8 and 12 which require
a padding of 0, 12, 8 and 4 respectively.