sysvinit: Fix out of bounds read
When dumping the ring buffer, the code checks to see if the previous
character was a newline, and if so starts a new line with a datestamp.
However this didn't take into account the buffer wrapping, and so when
the current position is the start of the buffer it would attempt to read
the character before the buffer, not the last character.