busybox: Update list of patches to remove
Scarthgap poky is still using busybox 1.36, but has added a number of
patches which are taken from the upstream busybox git repository.
However the CHERI git repository was forked from a later version
which already had these commits, so we need to remove them from the
list of patches to apply.

Signed-off-by: Stuart Menefy <stuart.menefy@codasip.com>