libgcrypt: Fixup inline assembly for CHERI
The library uses inline assembly code to prevent compiler optimisations.
In all cases there is no code in the assembly block, but registers are
passed in and out. Currently these registers are always use the "r"
constraint, but some of these are pointers (capabilities) and so need
to be modified to "C".

Signed-off-by: Stuart Menefy <stuart.menefy@codasip.com>